requirements.txt
alabaster==0.7.12
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreschema==0.0.4
Django==3.2.9
djangorestframework==3.12.4
docutils==0.18
drf-yasg==1.20.0
idna==3.3
imagesize==1.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
packaging==21.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
snowballstemmer==2.1.0
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.7
